requirements.txt
disnake
python-dotenv
deta